rtlbrowse: Split some function from stem_recurse.c to main.c
